fix ui & seeder & docker compose files

+++ backend/database/seeders/NetworkThingSeeder.php
@@ -22,7 +22,9 @@ class NetworkThingSeeder extends Seeder
     }
     public function run(): void
     {
-        $this->elasticsearchService->delete('network-things');
+        if($this->elasticsearchService->exist('network-things')) {
+            $this->elasticsearchService->delete('network-things');
+        }
         $this->elasticsearchService->create('network-things');
 
         DB::table('network_things')->truncate();
