+++ backend/database/seeders/ThingSeeder.php
@@ -31,7 +31,9 @@ class ThingSeeder extends Seeder
 
     public function run(): void
     {
-        $this->elasticsearchService->delete('things');
+        if ($this->elasticsearchService->exist('things')){
+            $this->elasticsearchService->delete('things');
+        }
         $this->elasticsearchService->create('things');
         //things-electronics
         DB::table('things')->truncate();
